Upload Base Files
Okay, here is our connection to the database. Some things to note...

You can put these files straight into your 'C://xampp/htdocs/' folder, and this should run just fine*
*provided you've created a database (using phpmyadmin [optional, just easiest]) called 'game' and have xampp running.

The way the db credentials are stored is NOT safe. If we eventually take this live, we would want to use the MVC design model as it would keep our db credentials outside of our root folder. For now, this will do; even if we took it live, I believe we could kinda protect accessing that file with an '.htaccess' file. Anyway, the person looking for this information would also need to know the file structure so that they could navigate to it in the browser. Regardless, we could make this more secure.